requirements.txt
torch
torchvision
# torchaudio
gensim 
scikit-learn
matplotlib
nltk
wandb
pyyaml
# librosa
pandas
umap-learn
# plotly
# kaleido